// src/app/core/model/MakePaymentModel.ts
export interface MakePaymentModel {
    mid: string;
    channelid: string;
    industrytypeid: string;
    website: string;
    custid: string;
    orderid: string;
    txnamount: string;
    callbackurl: string;
    checksum: string;
    paytmurl: string;
    mobileno: string;
    email: string;
}